Star Dust, A 2d Topdown Shooter
Created in over 4 months as a self learning experience. The main plan for this project was to learn how to create a  Quest system and Dialogue system in unity from scratch without outside help. The current project has 3 levels with different quests. This project was primarily made for android and was made so that it could run on low end android devices.